requirements.txt
langchain
python-dotenv
langchain-openai
langchain-community
langchain-core
bs4
faiss-cpu
streamlit
pypdf
tiktoken